Moves one e2e test to a new file, and adds several more
